All interactions now return an ActionOutput, which can be composed of more ActionOutputs if necessary. ActionResolver processes returned ActionOutputs into a single ActionResult, passing pending session updates through the chain.
This makes the code look a lot nicer, and also allows for pieces of output to change based on the results of previous output. Ie, when the giant chases you out of a room, the exits described will reflect the _new_ room rather than the one you were chased out of. Cool!